+++ emp-service-model/src/test/java/test/org/fugerit/java/emp/ex/TestWAExHelper.java
@@ -1,7 +1,9 @@
 package test.org.fugerit.java.emp.ex;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Response;
+import lombok.extern.slf4j.Slf4j;
 import org.fugerit.java.emp.ex.WAExHelper;
 import org.fugerit.java.emp.sm.service.ServiceMessage;
 import org.fugerit.java.emp.sm.service.ServiceResponse;
@@ -9,6 +11,10 @@ import org.fugerit.java.emp.sm.service.ServiceResponseHelper;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
+import java.io.IOException;
+import java.io.StringWriter;
+
+@Slf4j
 class TestWAExHelper {
 
     private static final String TEST_MESSAGE = "My Message";
@@ -27,6 +33,16 @@ class TestWAExHelper {
         Assertions.assertEquals( TEST_MESSAGE, this.getFirstError( WAExHelper.newEx(Response.Status.INTERNAL_SERVER_ERROR, TEST_MESSAGE ) ) );
         Assertions.assertEquals( TEST_MESSAGE, this.getFirstError( WAExHelper.newEx(Response.Status.INTERNAL_SERVER_ERROR, ServiceResponseHelper.newDefaultErrorMessage( TEST_MESSAGE ) ) ) );
         Assertions.assertEquals( TEST_MESSAGE, this.getFirstError( WAExHelper.newEx(Response.Status.INTERNAL_SERVER_ERROR, new ServiceResponse().addAllBySeverity( ServiceResponseHelper.newDefaultErrorMessage( TEST_MESSAGE ) ) ) ) );
+    }
+
+    @Test
+    void testSampleUsage() throws IOException {
+        WebApplicationException wae = WAExHelper.newEx( Response.Status.BAD_REQUEST, "Test bad request message" );
+        try (StringWriter writer = new StringWriter()) {
+            new ObjectMapper().writerWithDefaultPrettyPrinter().writeValue( writer, wae.getResponse() );
+            log.info( "test response : \n{}", writer );
+        }
+        Assertions.assertEquals( Response.Status.BAD_REQUEST.getStatusCode(), wae.getResponse().getStatus() );
     }
 
 }